% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biooracler-package.R
\docType{package}
\name{biooracler-package}
\alias{biooracler}
\alias{biooracler-package}
\title{biooracler: Import environmental ocean data from Bio-Oracle via ERDDAP}
\description{
Environmental ocean data modeled both for the present and for future climate change scenarios offered by the Bio-Oracle dataset \url{https://bio-oracle.org/} via ERDDAP server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bio-oracle.org/; https://erddap.bio-oracle.org/}
}

}
\author{
\strong{Maintainer}: Salvador Fernandez \email{salvador.fernandez@vliz.be}

Other contributors:
\itemize{
  \item Vinicius Salazar [contributor]
}

}
\keyword{internal}
