% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biooracler-package.R
\name{dataset_id}
\alias{dataset_id}
\title{Unique Identifier of a Bio-Oracle dataset in the ERDDAP Server}
\value{
Returns a help page
}
\description{
The \href{https://bio-oracle.org/}{Bio-Oracle} datasets have an unique identifier in the
\link{?ERDDAP} server where they are hosted.

Despite the title and summary of the dataset (See \code{\link[=list_layers]{list_layers()}}) already describe
the data included in the dataset, the \code{dataset_id} follow a certain naming that
indicate the information included inside. For instance:

-\code{phyc_baseline_2000_2020_depthsurf}
-\code{thetao_ssp119_2020_2100_depthmean}
-\code{thetao_baseline_2020_2019_depthsurf}
\subsection{Environmental variable}{

Indicated the first part of the naming (e.g. \code{phyc}, \code{thetao}). The equivalences are:
\itemize{
\item \code{chl} = Chlorophyll
\item \code{clt} = Total Cloud Fraction
\item \code{dfe} = Dissolved Iron
\item \code{mlotst} = Mixed Layer Depth
\item \code{no3} = Nitrate
\item \code{o2} = Dissolved Molecular Oxygen
\item \code{ph} = pH
\item \code{phyc} = Total Phytoplankton
\item \code{phyc} = Phytoplankton
\item \code{po4} = Phosphate
\item \code{si} = Silicate
\item \code{siconc} = Sea Ice Cover
\item \code{sithick} = Sea Ice Thickness
\item \code{so} = Salinity
\item \code{sws} = Sea Water Speed
\item \code{tas} = Air Temperature
\item \code{terrain} = Terrain
\item \code{thetao} = Ocean Temperature
}
}

\subsection{Time: Present/Past baseline or Future CMIIP6 Climate Change Projection projection}{

The part with \code{baseline} or \code{ssp119} indicates if the layer contains present or past
decades, or future climate change projections under the provisions of the
\href{https://www.ipcc.ch/}{Intergovernmental Panel on Climate Change} (IPCC) 6th report. This report features the
6th generation of \href{https://www.carbonbrief.org/qa-how-do-climate-models-work#cmip}{Coupled Model Intercomparison Projects} (CMIP6)
and describes a few \href{https://www.carbonbrief.org/explainer-how-shared-socioeconomic-pathways-explore-future-climate-change/}{Shared Socioeconomic Pathways} (SSPs), which are different
climate change scenarios.

The Bio-Oracle dataset predicts the environmental variables for 8 future decades
(2020-2100) and for 6 SSPs scenarios: \code{ssp119}, \code{ssp126}, \code{ssp245}, \code{ssp370}, \code{ssp460} and \code{ssp585}.

On the other hand, \code{baseline} refers to the conditions of the environmental variable
for the decades of 2000 to 2010 and 2010 to 2020, as interpolated from the
\href{https://marine.copernicus.eu/}{Copernicus Marine Service} data. It is named baseline
as this interpolated layers are the baseline to model the future predictions.
}

\subsection{Surface or Benthic}{

The Bio-Oracle datasets includes modelled environmental data for the surface sea water
layer (\code{depthsurf}), but also for the bottom of the ocean. Note that all datasets with
the subfixes \code{depthmin}, \code{depthmean} and \code{depthmax} refer to the benthic layer and
\emph{never refer to different depths of the ocean}. In other words: Bio-Oracle offers you
data from the ocean surface or the bottom, but \emph{never different levels of depth}.

The subfixes \code{min}, \code{mean} and \code{max} refer to the depth value used during modelling:
Bio-Oracle uses bathymetry data from \href{https://www.gebco.net/}{GEBCO}, which has a
higher resolution than the resolution of Bio-Oracle. Hence, for each grid cell of
the Bio-Oracle dataset there are more than one value of bathymetry Three different
bathymetry values were selected to fit the models: either the minimal value
(\code{depthmin}), an average of all the bathymetry values (\code{depthmean}) or the maximum
value (\code{depthmax}).
}
}
\examples{
\dontrun{?dataset_id}
}
