% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{download_dataset}
\alias{download_dataset}
\title{Downloads a griddap dataset from an ERDDAP server}
\usage{
download_dataset(
  dataset,
  variables,
  constraints,
  fmt = "nc",
  directory = FALSE,
  verbose = TRUE,
  debug = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Downloads a griddap dataset from an ERDDAP server
}
\examples{
\dontrun{
# Test variables
datasetid = "tas_baseline_2000_2020_depthsurf"
time = c('2001-01-01T00:00:00Z', '2010-01-01T00:00:00Z')
latitude = c(10, 20)
longitude = c(120, 130)
variables = c("tas_max", "tas_min")
constraints = list(time, latitude, longitude)
names(constraints) = c("time", "latitude", "longitude")

# Test call
download_dataset(datasetid, variables, constraints)
}
}
\seealso{
\link{list_layers}
}
