% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{download_griddap_dataset}
\alias{download_griddap_dataset}
\title{Downloads a griddap dataset from an ERDDAP server}
\usage{
download_griddap_dataset(
  erddap_server,
  dataset,
  variables,
  constraints,
  response = "nc",
  directory = "./",
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Downloads a griddap dataset from an ERDDAP server
}
\examples{

# Test variables
url = "https://erddap-test.emodnet.eu/erddap/"
datasetid = "biooracle_dmo_ds"
time = c('2001-01-01T00:00:00Z', '2010-01-01T00:00:00Z')
latitude = c(10, 20)
longitude = c(120, 130)
variables = c("o2_mean")
constraints = list(time, latitude, longitude)
names(constraints) = c("time", "latitude", "longitude")
# Test call
download_griddap_dataset(url, datasetid, variables, constraints)
}
