% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biooracler-package.R
\name{ERDDAP}
\alias{ERDDAP}
\title{Bio-Oracle ERDDAP Server}
\value{
Returns a help page
}
\description{
The \href{https://bio-oracle.org/}{Bio-Oracle} dataset is made available via an ERDDAP
server hosted by the \href{https://www.vliz.be}{Flanders Marine Institute (VLIZ)}.

ERDDAP is a data server that gives you simple, consistent way to download subsets
of scientific datasets in common file formats and make graphs and maps. The Bio-Oracle
ERDDAP installation is available online at:

http://erddap.bio-oracle.org/erddap/

ERDDAP offers a REST API that can be consumed by other software For instance,
there is an R package that reads data from ERDDAP named \link{rerddap}. This package is
the backbone behind \code{biooracler}: In fact, \code{biooracler} just wraps \link{rerddap} in a
convenient way to only read data from the Bio-Oracle ERDDAP server. But the Bio-Oracle
ERDDAP server is independent from the \code{biooracler} package.

You are welcome to access the Bio-Oracle ERDDAP server in your preferred way,
including via the \link{rerddap} R package. This will ease the combination of the
Bio-Oracle dataset with several other ERDDAP servers.

In addition, there is a Python extension that mimics the methods of \code{biooracler}. You
can access it at:

https://github.com/bio-oracle/pyo_oracle/

Each dataset in the Bio-Oracle ERDDAP server is identified by a \link{?dataset_id}.
}
\examples{
\dontrun{?ERDDAP}
}
