% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{info_layer}
\alias{info_layer}
\title{Gets detailed information about a single Bio-Oracle layer}
\usage{
info_layer(dataset_id)
}
\arguments{
\item{dataset_id}{An unique identifier of the dataset in \link{ERDDAP}. See \link{dataset_id}.}
}
\value{
An object of class \link{info} from the \link{rerddap} package
}
\description{
Gets detailed information about a single Bio-Oracle layer
}
\examples{
\dontrun{
info_layer("thetao_ssp119_2020_2100_depthmean")
}
}
\seealso{
\code{\link[=list_layers]{list_layers()}}, \code{\link[=download_layers]{download_layers()}}, \link{ERDDAP}, \link{dataset_id}
}
