% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{list_layers}
\alias{list_layers}
\title{List the layers in the Bio-Oracle dataset}
\usage{
list_layers(..., simplify = TRUE)
}
\arguments{
\item{...}{Free text search or pass params to \link[rerddap:ed_search_adv]{rerddap::ed_search_adv}}
}
\value{
a data frame with the dataset ID that can be later passed to \link{download_layers}
and further information
}
\description{
List the layers in the Bio-Oracle dataset
}
\examples{
\dontrun{
list_layers()
list_layers("Ocean Temperature 2100")
list_layers("thetao_ssp119_2020_2100_depthmean")
}
}
\seealso{
\link{download_layers}
}
